move clicked events to provider
