+++ packages/react-provider/src/constants/index.ts
@@ -1,0 +1,1 @@
+export const COURIER_CLIENT_HEADER = "x-courier-client-key";
