+++ packages/react-provider/src/hooks/use-actions.ts
@@ -1,8 +1,14 @@
+import { sendClickedRequest } from '~/lib/events';
 import useCourier from './use-courier';
 
 export default () => {
-  const { dispatch } = useCourier();
+  const { dispatch, clientKey } = useCourier();
   return {
+    handleOnClick: (data) => {
+      if (data?.clickedUrl) {
+        sendClickedRequest(clientKey, data?.clickedUrl);
+      }
+    },
     initToast: (payload) => {
       dispatch({
         type: "INIT_TOAST",
